Improved handling of repeats and focus of form controls
Handle setFocus (including in a repeat)

Set index after an insert or delete.

Recalculate rather than rebuild after a form change.

Separate "refresh" steps for outputs, repeats, and elements that use the index() function

Actions whose node set depends on an index now have the index calculated when the action is applied, not when the action is set. (This enables us to do a recalculate rather than a rebuild and hlpes with setting focus.)