Fix samples
Some samples used out of date paths to the current SEF.

I think I also accidentally deprecated the "initialTemplate" method of processing the XForm. I've fixed this and simplified the code that populates the $xforms-doc variable.

There was also a bug in the function library, which mangled XPaths to elements more than one level deep in the instance.