Update to handling of deferred updates and relevant fields
The "pending updates" behaviour has been replaced with an implementation of the XForms 1.1 "deferred update behaviour"

https://www.w3.org/TR/xforms11/#action-deferred-update-behavior

Reworked the relevant fields check, which now takes place as part of the refresh event.

The setvalue action now uses the recalculate mode. The "form-check" mode is now redundant and has been dropped.

Added "*:" prefix to templates for HTML elements (to allow XHTML handling).

Moved handling of "xforms-submit-done" actions to after the submission is competed.