Improvements to handling of bindings
A binding (xforms:bind) for a form control us now identified either by its @bind attribute or by matching the node with the nodeset returned by a <xforms:bind> element.

There is no longer any string comparison of XPath statements in the code.

The xforms-model-construct event now has a handler, and the xforms-rebuild event looks close to what is in the XForms 1.1 spec. A lot of code has been moved out of the xformsjs-main template.

Handling of actions and calculations is simpler (by identifying the instance context when the form is built), with fewer XSLT template and function calls.

The Javascript section is removed to a separate imported stylesheet (for my ease of navigation)